+++ app/events/page.tsx
@@ -1,6 +1,6 @@
 "use client";
-import Footer from "@/components/footer";
-import Header from "@/components/header";
+import Footer from "@/components/_footer";
+import Header from "@/components/_header";
 import EventsCard from "@/components/organization/events_card";
 import { createClient, getUser } from "@/lib/supabase/client";
 import { Event } from "@/types/event";
