+++ app/user/profile/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Footer from "@/components/footer";
-import Header from "@/components/header";
+import Footer from "@/components/_footer";
+import Header from "@/components/_header";
 import UserEvents from "@/components/user/user_events";
 import UserOrganizations from "@/components/user/user_organizations";
 import { fetchEventsForUser } from "@/lib/events";
